requirements.txt
requests
BeautifulSoup4
aiogram
psycopg2-binary
peewee
python-decouple